+++ example_shared_ubwa_instance.py
@@ -50,7 +50,7 @@ exchange = "binance.com"
 
 ubwa = BinanceWebSocketApiManager(exchange=exchange)
 ubldc = BinanceLocalDepthCacheManager(exchange=exchange, ubwa_manager=ubwa)
-ubldc.create_depth_cache(market=market)
+ubldc.create_depth_cache(markets=market)
 
 while True:
     try: